fix(googlechat): align channel conventions with existing implementations
- Add default GroupPolicy "pairing" for DB instances (security parity with Discord/Feishu/Slack)
- Fix BaseChannel name to use channels.TypeGoogleChat constant ("google_chat")
- Warn at startup when OIDC verification is disabled (no project_number)
- Extract resolveServiceAccountJSON to credentials.go (channel.go under 200 lines)
